+++ openvino_xai/metrics/__init__.py
@@ -1,0 +1,5 @@
+# Copyright (C) 2024 Intel Corporation
+# SPDX-License-Identifier: Apache-2.0
+"""
+Metrics in OpenVINO-XAI to check the quality of saliency maps.
+"""